Move new dashboard header into shared components
