+++ src/app/shared/components/new-dashboard-header/dashboard-header.component.scss
